// apps/web/src/lib/generated/taxonomy/taxonomy-options.gen.ts
export const erasOptions: readonly SelectOption[] = [
  { value: 'jahili', label: 'جاهلي', poemsCount: 7715 },
  { value: 'islami', label: 'إسلامي', poemsCount: 9777 },
  { value: 'umawi', label: 'أموي', poemsCount: 14335 },
  { value: 'abbasi', label: 'عباسي', poemsCount: 39667 },
  { value: 'andalusi', label: 'أندلسي', poemsCount: 12642 },
  { value: 'fatimi', label: 'فاطمي', poemsCount: 11740 },
  { value: 'ayyubi', label: 'أيوبي', poemsCount: 10215 },
  { value: 'mamluki', label: 'مملوكي', poemsCount: 17667 },
  { value: 'uthmani', label: 'عثماني', poemsCount: 13596 },
  { value: 'hadith', label: 'حديث', poemsCount: 188507 },
  { value: 'muasir', label: 'معاصر', poemsCount: 3471 },
  { value: 'ghayrmaruf', label: 'غير معروف', poemsCount: 51408 },
];